refactor: move authentication-related controllers to a single folder
