+++ src/main/factories/controllers/login/LoginControllerFactory.ts
@@ -1,7 +1,7 @@
 import LogControllerDecorator from '../../../decorators/LogControllerDecorator'
 import Controller from '../../../../presentation/protocols/Controller'
 import LogMongoRepository from '../../../../infra/db/mongodb/log-repository/LogMongoRepository'
-import LoginController from '../../../../presentation/controllers/login/LoginController'
+import LoginController from '../../../../presentation/controllers/authentication/login/LoginController'
 import LoginValidationCompositeFactory from './LoginValidationCompositeFactory'
 import DbAuthenticateUserFactory from '../../usecases/authenticate-user/DbAuthenticateUserFactory'
