+++ src/presentation/controllers/authentication/login/LoginController.ts
@@ -1,9 +1,9 @@
-import Controller from '../../protocols/Controller'
-import HttpRequest from '../../protocols/HttpRequest'
-import HttpResponse from '../../protocols/HttpResponse'
-import HttpHelper from '../../helpers/http/HttpHelper'
-import AuthenticateUser from '../../../domain/usecases/AuthenticateUser'
-import Validation from '../../protocols/Validation'
+import Controller from '../../../protocols/Controller'
+import HttpRequest from '../../../protocols/HttpRequest'
+import HttpResponse from '../../../protocols/HttpResponse'
+import HttpHelper from '../../../helpers/http/HttpHelper'
+import AuthenticateUser from '../../../../domain/usecases/AuthenticateUser'
+import Validation from '../../../protocols/Validation'
 
 export default class LoginController implements Controller {
   constructor (
